updating file names and cleaning up
